feat: 1119 Certificate in debug...
